+++ route_breakdown.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
+	"errors"
 	"fmt"
 	"net/http"
 	"sync"
@@ -128,6 +129,15 @@ func (s *routeBreakdowns) send(m map[routeBreakdownKey]*routeBreakdown) error {
 	switch resp.StatusCode {
 	case http.StatusUnauthorized:
 		return errUnauthorized
+	case http.StatusTooManyRequests:
+		return errIPRateLimited
+	case http.StatusBadRequest:
+		var sendResp sendResponse
+		err = json.NewDecoder(buf).Decode(&sendResp)
+		if err != nil {
+			return err
+		}
+		return errors.New(sendResp.Message)
 	}
 
 	err = fmt.Errorf("got unexpected response status=%q", resp.Status)
